requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlmodel
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
email-validator
